feat : bemybaby 좋아요 기능 추가, 피드 Get return값 댓글, 좋아요 포함하도록 수정

+++ woochuchu/bemybaby/views.py
@@ -1,4 +1,3 @@
-from datetime import date
 from django.shortcuts import render, get_object_or_404
 from rest_framework.response import Response
 from rest_framework import status
@@ -7,22 +6,53 @@ from .serializers import *
 from .models import *
 from drf_yasg.utils import swagger_auto_schema
 from drf_yasg import openapi
+from accounts.permissions import *
 
 # 피드도 S3 때문에 커스터마이징 위해서 APIView 이용해서 하는 걸로 수정
 class BeMyBabyAPIView(APIView):
-    def get_objects(self):
+    permission_classes = [
+        JwtPermission.IsAuthenticatedOrReadOnly
+    ]
+
+    def get_feed_objects(self):
         return BeMyBaby.objects.all().order_by('-id')
 
+    def get_comment_objects(self, feed_id):
+        return BeMyBabyComment.objects.filter(bemybaby_id=feed_id).order_by('id')
+
     def get(self, request):
+        """
+        피드를 조회합니다.
+        """
         try :
-            feeds = self.get_objects()
-            serializer = BeMyBabySerializer(feeds, many=True)
+            comment_paired_feeds = []
+            feeds = self.get_feed_objects()
+            for feed in feeds:
+                #피드 본문 처리
+                feed_serializer = BeMyBabySerializer(feed)
+                #피드 댓글 처리
+                comments = self.get_comment_objects(feed.id)
+                comment_serializer = BeMyBabyCommentSerializer(comments, many=True)
+                comment = {
+                    "comments": comment_serializer.data
+                }
+                #피드 좋아요 처리
+                likes = BeMyBabyLike.objects.filter(bemybaby_id=feed.id)
+                likes_count = likes.count()
+                data = feed_serializer.data
+                likes = {
+                    "likes_count": likes_count
+                }
+                data.update(comment)
+                data.update(likes)
+                comment_paired_feeds.append(data)
             data = {
                 "results": {
-                    "data": serializer.data
+                    "data": comment_paired_feeds
                 }
             }
             return Response(data=data, status=status.HTTP_200_OK)
+
         except Exception as e:
             print(e)
             data = {
@@ -34,7 +64,11 @@ class BeMyBabyAPIView(APIView):
             return Response(data=data, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
     def post(self, request):
+        """
+        새 피드를 작성합니다.
+        """
         try:
+            request.data['user'] = request.user.id
             serializer = BeMyBabySerializer(data=request.data)
             if serializer.is_valid():
                 serializer.save()
@@ -44,6 +78,7 @@ class BeMyBabyAPIView(APIView):
                     }
                 }
                 return Response(data=data, status=status.HTTP_200_OK ) 
+
             else:
                 data = {
                     "results": {
@@ -52,6 +87,7 @@ class BeMyBabyAPIView(APIView):
                     }
                 }
                 return Response(data=data, status=status.HTTP_400_BAD_REQUEST)
+
         except Exception as e:
             print(e)
             data = {
@@ -63,29 +99,48 @@ class BeMyBabyAPIView(APIView):
             return Response(data=data, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 class BeMyBabyDeletePutView(APIView):
+    permission_classes = [
+        JwtPermission.IsAuthorUpdateDeleteorReadOnly
+    ]
     def get_object(self, feed_id):
         return BeMyBaby.objects.get(id=feed_id)
 
     def put(self, request, feed_id):
+        """
+        피드를 수정합니다.
+        """
         try :
             feed = self.get_object(feed_id=feed_id)
-            serializer = BeMyBabySerializer(feed, data=request.data)
-            if serializer.is_valid():
-                serializer.save()
+            if request.user.id != feed.user.id:
                 data = {
                     "results": {
-                        "msg": "데이터가 성공적으로 저장되었습니다."
+                        "msg": "권한이 없습니다." 
                     }
                 }
-                return Response(data=data, status=status.HTTP_200_OK)
+
+                return Response(data=data, status=status.HTTP_401_UNAUTHORIZED)
+
             else:
-                data = {
-                    "results": {
-                        "msg": serializer.errors,
-                        "code": "E4000"
+                request.data['user'] = request.user.id
+                serializer = BeMyBabySerializer(feed, data=request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    data = {
+                        "results": {
+                            "msg": "데이터가 성공적으로 저장되었습니다."
+                        }
                     }
-                }
-                return Response(data=data, status=status.HTTP_400_BAD_REQUEST)
+                    return Response(data=data, status=status.HTTP_200_OK)
+
+                else:
+                    data = {
+                        "results": {
+                            "msg": serializer.errors,
+                            "code": "E4000"
+                        }
+                    }
+                    return Response(data=data, status=status.HTTP_400_BAD_REQUEST)
+
         except BeMyBaby.DoesNotExist:
             data = {
                 "results": {
@@ -94,6 +149,7 @@ class BeMyBabyDeletePutView(APIView):
                 }
             }
             return Response(data=data, status=status.HTTP_404_NOT_FOUND)
+
         except Exception as e:
             print(e)
             data = {
@@ -105,15 +161,27 @@ class BeMyBabyDeletePutView(APIView):
             return Response(data=data, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
     def delete(self, request, feed_id):
+        """
+        피드를 삭제합니다.
+        """
         try:
             feed = self.get_object(feed_id=feed_id)
-            feed.delete()
-            data = {
-                "results": {
-                    "msg": "데이터가 성공적으로 삭제되었습니다."
+            if request.user.id != feed.user.id:
+                data = {
+                    "results": {
+                        "msg": "권한이 없습니다." 
+                    }
                 }
-            }
-            return Response(data=data, status=status.HTTP_200_OK)
+                return Response(data=data, status=status.HTTP_401_UNAUTHORIZED)
+            else:
+                feed.delete()
+                data = {
+                    "results": {
+                        "msg": "데이터가 성공적으로 삭제되었습니다."
+                    }
+                }
+                return Response(data=data, status=status.HTTP_200_OK)
+
         except BeMyBaby.DoesNotExist:
             data = {
                 "results": {
@@ -122,6 +190,7 @@ class BeMyBabyDeletePutView(APIView):
                 }
             }
             return Response(data=data, status=status.HTTP_404_NOT_FOUND)
+
         except Exception as e:
             print(e)
             data = {
@@ -133,10 +202,17 @@ class BeMyBabyDeletePutView(APIView):
             return Response(status=status.HTTP_404_NOT_FOUND)
 
 class BeMyBabyCommentAPIView(APIView):
-    def get_objects(self, feed_id):
-        return BeMyBabyComment.objects.filter(bemybaby_id=feed_id)
+    permission_classes = [
+        JwtPermission.IsAuthenticatedOrReadOnly
+    ]
 
+    def get_objects(self, feed_id):
+        return BeMyBabyComment.objects.filter(bemybaby_id=feed_id).order_by('id')
+    '''
     def get(self, request, feed_id):
+        """
+        특정 피드의 댓글들을 조회합니다.
+        """
         try:
             comments = self.get_objects(feed_id=feed_id)
             serializer = BeMyBabyCommentSerializer(comments, many=True)
@@ -146,6 +222,7 @@ class BeMyBabyCommentAPIView(APIView):
                 }
             }
             return Response(data=data, status=status.HTTP_200_OK)
+
         except Exception as e:
             print(e)
             data = {
@@ -154,13 +231,18 @@ class BeMyBabyCommentAPIView(APIView):
                     "code": "E5000"
                 }
             }
-            return Response(data=data, status=status.HTTP_400_BAD_REQUEST)
-    
+            return Response(data=data, status=status.HTTP_500_BAD_REQUEST)
+    '''
     def post(self, request, feed_id):
+        """
+        특정 피드에 댓글을 작성합니다.
+        """
         try:
             request.data['bemybaby'] = feed_id
+            request.data['user'] = request.user.id
             serializer = BeMyBabyCommentSerializer(data=request.data)
             if serializer.is_valid():
+                serializer.bemybaby_id = feed_id
                 serializer.save()
                 data = {
                     "results": {
@@ -168,6 +250,7 @@ class BeMyBabyCommentAPIView(APIView):
                     }
                 }
                 return Response(data=data, status=status.HTTP_200_OK ) 
+
             else:
                 data = {
                     "results": {
@@ -176,6 +259,7 @@ class BeMyBabyCommentAPIView(APIView):
                     }
                 }
                 return Response(data=data, status=status.HTTP_400_BAD_REQUEST)
+
         except Exception as e:
             print(e)
             data = {
@@ -184,22 +268,51 @@ class BeMyBabyCommentAPIView(APIView):
                     "code": "E5000"
                 }
             }
-            return Response(status=status.HTTP_400_BAD_REQUEST)
+            return Response(data=data, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 class BeMyBabyCommentDeletePutAPIView(APIView):
+    permission_classes = [
+        JwtPermission.IsAuthorUpdateDeleteorReadOnly
+    ]
+
     def get_object(self, comment_id):
         return BeMyBabyComment.objects.get(id=comment_id)
 
-    def delete(self, request, comment_id):
+    def put(self, request, comment_id):
+        """
+        댓글을 수정합니다.
+        """
         try:
             comment = self.get_object(comment_id=comment_id)
-            comment.delete()
-            data = {
-                "results": {
-                    "msg": "데이터가 성공적으로 삭제되었습니다."
+            request.data['bemybaby'] = comment.bemybaby_id
+            if request.user.id != comment.user.id:
+                data = {
+                    "results": {
+                        "msg": "권한이 없습니다." 
+                    }
                 }
-            }
-            return Response(data=data, status=status.HTTP_200_OK)
+
+                return Response(data=data, status=status.HTTP_401_UNAUTHORIZED)
+            else:
+                request.data['user'] = request.user.id
+                serializer = BeMyBabyCommentSerializer(comment, data=request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    data = {
+                        "results": {
+                            "msg": "데이터가 성공적으로 저장되었습니다."
+                        }
+                    }
+                    return Response(data=data, status=status.HTTP_200_OK) 
+
+                else:
+                    data = {
+                        "results": {
+                            "msg": serializer.errors,
+                            "code": "E4000"
+                        }
+                    }
+                    return Response(data=data, status=status.HTTP_400_BAD_REQUEST)
 
         except BeMyBabyComment.DoesNotExist:
             data = {
@@ -218,28 +331,41 @@ class BeMyBabyCommentDeletePutAPIView(APIView):
                     "code": "E5000"
                 }
             }
-            return Response(data=data, status=status.HTTP_404_NOT_FOUND)
+            return Response(data=data, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
-    def put(self, request, comment_id):
+    def delete(self, request, comment_id):
+        """
+        댓글을 삭제합니다.
+        """
         try:
             comment = self.get_object(comment_id=comment_id)
-            serializer = BeMyBabyCommentSerializer(comment, data=request.data)
-            if serializer.is_valid():
-                serializer.save()
+            if request.user.id != comment.user.id:
                 data = {
                     "results": {
-                        "msg": "데이터가 성공적으로 저장되었습니다."
+                        "msg": "권한이 없습니다." 
                     }
                 }
-                return Response(data=data, status=status.HTTP_200_OK) 
+
+                return Response(data=data, status=status.HTTP_401_UNAUTHORIZED)
+
             else:
+                comment.delete()
                 data = {
                     "results": {
-                        "msg": serializer.errors,
-                        "code": "E4000"
+                        "msg": "데이터가 성공적으로 삭제되었습니다."
                     }
                 }
-                return Response(data=data, status=status.HTTP_400_BAD_REQUEST)
+                return Response(data=data, status=status.HTTP_200_OK)
+
+        except BeMyBabyComment.DoesNotExist:
+            data = {
+                "results": {
+                    "msg": "일치하는 데이터가 존재하지 않습니다.",
+                    "code": "E4040"
+                }
+            }
+            return Response(data=data, status=status.HTTP_404_NOT_FOUND)
+
         except Exception as e:
             print(e)
             data = {
@@ -250,3 +376,43 @@ class BeMyBabyCommentDeletePutAPIView(APIView):
             }
             return Response(data=data, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
+class BeMyBabyLikeAPIView(APIView):
+    permission_classes = [
+        JwtPermission.IsAuthorUpdateDeleteorReadOnly
+    ]
+
+    def get_objects(self, user, feed_id):
+        return BeMyBabyLike.objects.get(bemybaby_id=feed_id, user_id=user.id)
+
+    def post(self, request, feed_id):
+        try: 
+            like = self.get_objects(request.user, feed_id)
+            like.delete()
+            like_count = BeMyBabyLike.objects.filter(bemybaby_id=feed_id).count()
+            data = {
+                "results": {
+                    "data": like_count 
+                }
+            }
+            return Response(data=data, status=status.HTTP_200_OK)
+
+        except BeMyBabyLike.DoesNotExist:
+            BeMyBabyLike.objects.create(user=request.user, bemybaby_id=feed_id)
+            like_count = BeMyBabyLike.objects.filter(bemybaby_id=feed_id).count()
+            data = {
+                "results": {
+                    "data": like_count 
+                }
+            }
+            return Response(data=data, status=status.HTTP_200_OK)
+
+        except Exception as e:
+            print(e)
+            data = {
+                "results": {
+                    "msg": "정상적인 접근이 아닙니다.",
+                    "code": "E5000"
+                }
+            }
+            
+            return Response(data=data, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
